ai_agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_agent"
version = "0.1.0"
description = "LangGraph-based AI agent service for LoadRouter"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "langgraph>=0.0.40",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "psycopg2-binary>=2.9",
  "python-dotenv>=1.0",
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "openai>=1.12",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ai_agent = ["prompts/*.txt", "prompts/*.json"]
